Refactor project structure and enhance development workflow
- Add comprehensive TypeScript configuration with strict type checking
- Implement build system with proper output directory structure
- Add development tools: ESLint, Prettier, Jest for testing
- Update package.json with complete scripts and metadata
- Relocate instrumentation.ts to src/ directory for better organization
- Add dist/ and BLOG.md to .gitignore
- Update dependencies to latest versions
