+++ package.json
@@ -1,36 +1,52 @@
 {
   "name": "aie-feb-25-starter",
   "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
+  "description": "A production-ready Node.js service.",
+  "main": "dist/index.js",
+  "type": "commonjs",
+  "files": [
+    "dist"
+  ],
   "scripts": {
-    "dev": "tsx -r ./instrumentation.ts src/index.ts"
+    "dev": "tsx watch --clear-screen=false ./src/index.ts",
+    "build": "tsc --build",
+    "start": "node dist/index.js",
+    "lint": "eslint . --ext .ts --fix",
+    "format": "prettier --write .",
+    "test": "jest"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
     "@ai-sdk/google": "^1.2.22",
-    "@ai-sdk/openai": "^1.3.22",
+    "@ai-sdk/openai": "^1.3.23",
     "@ai-sdk/perplexity": "^1.1.9",
-    "@effect/platform": "^0.87.9",
-    "@effect/platform-node": "^0.89.1",
-    "@opentelemetry/auto-instrumentations-node": "^0.60.1",
+    "@effect/platform": "^0.87.11",
+    "@effect/platform-node": "^0.89.3",
+    "@opentelemetry/auto-instrumentations-node": "^0.61.0",
     "@opentelemetry/sdk-node": "^0.202.0",
     "@opentelemetry/sdk-trace-node": "^2.0.1",
     "@t3-oss/env-core": "^0.13.8",
-    "ai": "^4.3.16",
+    "ai": "^4.3.17",
     "effect": "^3.16.12",
     "exa-js": "^1.8.19",
-    "langfuse": "^3.38.2",
-    "langfuse-vercel": "^3.38.2",
+    "langfuse": "^3.38.4",
+    "langfuse-vercel": "^3.38.4",
     "ollama-ai-provider": "^1.2.0",
-    "zod": "^3.25.74"
+    "zod": "^3.25.76"
   },
   "devDependencies": {
-    "@types/node": "^24.0.10",
-    "dotenv": "^17.0.1",
-    "tsx": "^4.20.3"
+    "@types/jest": "^30.0.0",
+    "@types/node": "^24.0.12",
+    "dotenv": "^17.1.0",
+    "eslint": "^9.30.1",
+    "eslint-config-prettier": "^10.1.5",
+    "jest": "^30.0.4",
+    "prettier": "^3.6.2",
+    "ts-jest": "^29.4.0",
+    "tsx": "^4.20.3",
+    "typescript": "^5.8.3"
   },
   "packageManager": "pnpm@10.12.1+sha512.f0dda8580f0ee9481c5c79a1d927b9164f2c478e90992ad268bbb2465a736984391d6333d2c327913578b2804af33474ca554ba29c04a8b13060a717675ae3ac"
 }
