+++ src/server.ts
@@ -1,5 +1,5 @@
 import { Layer } from "effect";
-import { HttpApiBuilder, HttpApiMiddleware, HttpApiSwagger } from "@effect/platform";
+import { HttpApiBuilder, HttpApiSwagger } from "@effect/platform";
 import { NodeHttpServer, NodeRuntime } from "@effect/platform-node";
 import { createServer } from "node:http";
 import { DeepResearchApiLive } from "./handlers/deep-research";
