+++ src/services/utils.ts
@@ -1,6 +1,6 @@
 import { generateObject as aiGenerateObject, generateText as aiGenerateText } from "ai";
 import { Data, Effect } from "effect";
-import { ERROR_TYPES, OPENAI_MODELS, Research } from "../constants";
+import { ERROR_TYPES, Research } from "../constants";
 import { ZodSchema } from "zod";
 import { AiModels } from "./AiModels";
 import { randomUUID } from "crypto";
@@ -29,12 +29,19 @@ class GenerateTextError extends Data.TaggedError(ERROR_TYPES.GENERATE_TEXT_ERROR
 
 type GenerateObjectOptions = Parameters<typeof aiGenerateObject>[0]
 type GenerateTextOptions = Parameters<typeof aiGenerateText>[0];
-type Options =
+type ObjectOptions =
 	// start by dropping `output`, `model` and `schema` all at once
 	Omit<GenerateObjectOptions, 'output' | 'model'>
 	// now re‐add `model` and `schema`, but make them optional
 	& Partial<Pick<GenerateObjectOptions, 'model'>>
 	;
+type TextOptions =
+	// start by dropping `output`, `model` and `schema` all at once
+	Omit<GenerateTextOptions, 'output' | 'model'>
+	// now re‐add `model` and `schema`, but make them optional
+	& Partial<Pick<GenerateTextOptions, 'model'>>
+	;
+
 
 interface TraceMetadata {
 	calledFrom: string;
@@ -44,13 +51,13 @@ interface TraceMetadata {
 
 
 
-export const generateObject = Effect.fn("generateObject")(function* (options: Options, schema: ZodSchema, metadata: TraceMetadata) {
+export const generateObject = Effect.fn("generateObject")(function* (options: ObjectOptions, schema: ZodSchema, metadata: TraceMetadata) {
 	const { openai } = yield* AiModels
 	return yield* Effect.tryPromise({
 		try: () => {
 
 			const res = aiGenerateObject({
-				...options,
+
 				model: openai('gpt-4.1-mini'),
 				output: 'object',
 				schema,
@@ -63,6 +70,7 @@ export const generateObject = Effect.fn("generateObject")(function* (options: Op
 						langfuseUpdateParent: false
 					}
 				},
+				...options,
 			})
 
 			return res
@@ -73,12 +81,14 @@ export const generateObject = Effect.fn("generateObject")(function* (options: Op
 })
 
 
-export const generateEnum = Effect.fn("generateEnum")(function* (options: Omit<GenerateObjectOptions, 'output'>, enums: string[], metadata: TraceMetadata) {
+export const generateEnum = Effect.fn("generateEnum")(function* (options: ObjectOptions, enums: string[], metadata: TraceMetadata) {
+	const { openai } = yield* AiModels
 	return yield* Effect.tryPromise({
 		try: () => {
 
 			const res = aiGenerateObject({
-				...options,
+
+				model: openai('gpt-4.1-mini'),
 				output: 'enum',
 				enum: enums,
 				experimental_telemetry: {
@@ -90,6 +100,7 @@ export const generateEnum = Effect.fn("generateEnum")(function* (options: Omit<G
 						langfuseUpdateParent: false
 					}
 				},
+				...options,
 			})
 
 
@@ -101,12 +112,13 @@ export const generateEnum = Effect.fn("generateEnum")(function* (options: Omit<G
 })
 
 
-export const generateText = Effect.fn("generateText")(function* (options: GenerateTextOptions, metadata: TraceMetadata) {
+export const generateText = Effect.fn("generateText")(function* (options: TextOptions, metadata: TraceMetadata) {
+	const { openai } = yield* AiModels;
 	return yield* Effect.tryPromise({
 		try: () => {
 
 			const res = aiGenerateText({
-				...options,
+				model: openai('gpt-4.1-mini'),
 				experimental_telemetry: {
 					isEnabled: true,
 					functionId: `generate-text-function-${randomUUID()}`,
@@ -116,6 +128,7 @@ export const generateText = Effect.fn("generateText")(function* (options: Genera
 						langfuseUpdateParent: false
 					}
 				},
+				...options,
 			})
