+++ tsconfig.json
@@ -1,0 +1,38 @@
+{
+  "compilerOptions": {
+    // Basic
+    "target": "ES2022", // modern runtime
+    "module": "NodeNext", // for Node.js ≤ 18; switch to "ESNext" + "type":"module" if desired
+    "rootDir": "src", // your source root
+    "outDir": "dist", // emit compiled files here
+    "declaration": true, // generate .d.ts for published packages
+    "sourceMap": true, // map for debugging in prod
+    // Strictness
+    "strict": true,
+    "noImplicitAny": true,
+    "noImplicitThis": true,
+    "alwaysStrict": true,
+    "noUnusedLocals": true,
+    "noUnusedParameters": true,
+    "noUncheckedIndexedAccess": true,
+    "noImplicitOverride": true,
+    // Module resolution & interop
+    "esModuleInterop": true,
+    "forceConsistentCasingInFileNames": true,
+    "resolveJsonModule": true,
+    // Speed
+    "skipLibCheck": true,
+    // Isolated modules & tools
+    "isolatedModules": true, // required by tsx / babel
+    "moduleResolution": "NodeNext",
+    "allowJs": false // typically false in prod TS codebase
+  },
+  "include": [
+    "src/**/*"
+  ],
+  "exclude": [
+    "node_modules",
+    "dist",
+    "**/*.test.ts"
+  ]
+}
